Split DWARFImporterDelegate into two parts. NFC
1. SwiftDWARFImporterForClangTypes is owned by a TypeSystemSwiftTypeRef.
   It imports Clang types from DWARF.

2. SwiftDWARFImporterDelegate is a swift::DWARFImporterDelegate that uses
   clang::ASTImporter to import a type from the TypeSystemClang used by
   the DWARFParser into the clang ASTContext used by one SwiftASTContext's
   ClangImporter.

This is in preparation for creating more flexible SwiftASTContexts in the future.

(cherry picked from commit 47f25ae8c7c07da1d89e6467d9d7d7ba644d47c1)
